requirements.txt
playwright==1.44.0
python-dotenv==1.0.0
aiohttp==3.9.5
